REFACTOR (IN PROG(
Due to the increasing amounts of simulations, a big jounrey of refactoring has begun to treat all the simulations as the same object and to adhere to the DRY principal. currently 1 of 5 sims has been refactored into this new system, but a really good basis has been created for the rest of the sims to sweamlessly transition into!